Fix missing SortDescription namespace
