+++ Enums.cs
@@ -1,0 +1,6 @@
+namespace BinanceUsdtTicker
+{
+    public enum FilterMode { All, Positive, Negative }
+    public enum ThemeKind { Light, Dark }
+    public enum QuickFilter { None, Pos3Plus, Neg3Minus }
+}
