+++ Models/ColumnState.cs
@@ -1,0 +1,9 @@
+namespace BinanceUsdtTicker.Models
+{
+    public class ColumnState
+    {
+        public string Header { get; set; } = "";
+        public int DisplayIndex { get; set; }
+        public double Width { get; set; }
+    }
+}
